Not Initial But Initial Commit XD
I didn't know where i initialized the local repository until 1:05 PM so i didn't push anything for about 7 hours. This code works 89% of the first stage. Please don't forget that an.py is the functional one. The others are tests. an.py was born after selenium and bs4 had married.